requirements.txt
fastapi
uvicorn
torch
facenet-pytorch
pillow
psycopg2
redis
pgvector
pytest
pytest-asyncio
pydantic
asyncio
numpy